requirements.txt
requests>=2.31.0
py-clob-client>=0.23.0
websocket-client>=1.7.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
